fix many typos, move transform to utils
